+++ CHANGELOG.md
@@ -7,3 +7,7 @@
 -   Example project names to be more realistic
 
 -   **BREAKING CHANGE**: `--stage` flag to `--env`, functionality has not changed.
+
+-   **BREAKING CHANGE**: `<project_path>` input option to a `--cwd` flag for all CLI commands
+
+-   **BREAKING CHANGE**: `bm server scope --project <project>` now sets project using `bm server scope <project>`